% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{get_params}
\alias{get_params}
\title{Get default model parameters}
\usage{
get_params(
  beta_r = 0.1,
  beta_s = 0.2,
  gamma_r = 0.05,
  gamma_s = 0.1,
  omega = 1/(1 * 365),
  lsr = 0.02,
  lrs = 0.01,
  delta_sr = 0.005,
  delta_rs = 0.005,
  c0 = 0.5,
  theta = 0.01,
  lambda = 0.005,
  rho = 0.2,
  mu = 1/(50 * 365),
  N = 1000
)
}
\arguments{
\item{beta_r}{Infection rate from mosquitoes to resistant humans}

\item{beta_s}{Infection rate from mosquitoes to sensitive humans}

\item{gamma_r}{Recovery rate for resistant humans}

\item{gamma_s}{Recovery rate for sensitive humans}

\item{omega}{Waning immunity}

\item{lsr}{Rate of change from sensitive to resistant}

\item{lrs}{Rate of change from resistant to sensitive}

\item{delta_sr}{Rate of change from sensitive infected to resistant infected}

\item{delta_rs}{Rate of change from resistant infected to sensitive infected}

\item{c0}{Initial ITN coverage}

\item{theta}{ITN decay rate}

\item{lambda}{ITN replenishment rate}

\item{rho}{Fraction of recovered returning to Sr}

\item{mu}{Birth/Death rate}

\item{N}{Total human population}
}
\description{
This function returns the default parameter values for the malaria model.
\code{get_params} specifies model default parameters
}
