% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packmalaria-package.R
\docType{package}
\name{packmalaria-package}
\alias{packmalaria}
\alias{packmalaria-package}
\title{packmalaria: Infectious disease modelling and simulation}
\description{
Packmalaria is mechanistic model that aims to achieve simuation of realistic scenarios of malaria transmission. It presently incorporates human resistance and sensitivity in the modelling the disease transmission.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Olatundemi/packmalaria}
  \item Report bugs at \url{https://github.com/Olatundemi/packmalaria/issues}
}

}
\author{
\strong{Maintainer}: Olatunde Ibrahim \email{o.ibrahim@imperial.ac.uk}

}
\keyword{internal}
