% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_model.R
\name{simulate_model}
\alias{simulate_model}
\title{Function that return deterministic/stochastic simulation of the model}
\usage{
simulate_model(
  gen = NULL,
  params = get_params(),
  time_steps = 365,
  n_particles = 1,
  output_format = "dataframe"
)
}
\arguments{
\item{gen}{compiled model (default).}

\item{params}{User's defined parameters for model compilation.}

\item{time_steps}{duration of simulation times usually expressed as a measure of days}

\item{n_particles}{number of particles to determine model stochasticity}

\item{output_format}{simulation output format}
}
\description{
\code{simulate_model} # A simulation function that creates a dust2 system from the SIR model.
}
