% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packmalaria-package.R
\docType{package}
\name{packmalaria-package}
\alias{packmalaria}
\alias{packmalaria-package}
\title{packmalaria: Infectious disease modelling and simulation}
\description{
Packmalaria aims to achieve simuation of realistic scenarios of malaria transmission
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Olatundemi/packmalaria}
  \item Report bugs at \url{https://github.com/Olatundemi/packmalaria/issues}
}

}
\author{
\strong{Maintainer}: Olatunde Ibrahim \email{oibrahim@imperial.ac.uk}

}
\keyword{internal}
